feat: Add initial support for handlers (#217)
Related #214 


- Add the handlers to the graph with `--show-handlers`. **This is the
initial support for handlers.** They are by default added at the end of
the play and roles only. This doesn't reflect Ansible behavior.
- Changes the shape of the graphviz node to make it consistent with
Mermaid. The tasks will be rectangle instead of `octagon`:
https://graphviz.org/doc/info/shapes.html
- Refactor how the node/task index are computed given we can now add
handlers after all the tasks are parsed.
- Add a new `display_name()` method to the node for a friendly name for
the graph. This removes the of passing the `node_label_prefix` in
multiple places.
- Remove the class **CompositeTasksNode** as it is no longer needed
anymore.
- Remove the play name from the edge going from playbook to the plays.
This was not consistent with the other edges.